Mudanças na Renderer concluídas.
* Adicionado um char teste,
* Algumas correções feitas no modo como a
EventQueue trata eventos de teclado

+++ ConsoleApplication1/ConsoleApplication1/Renderer.hpp
@@ -15,7 +15,7 @@ private:
 	int _screenw, _screenh; //Tamanho da tela
 
 	int _screenx, _screeny; //Onde a tela está (em relação ao jogo)
-	int _unitsize; //Quantos pixels são necessários para 1 unidade no jogo?
+	float _unitsize; //Quantos pixels são necessários para 1 unidade no jogo?
 	
 	CharacterManager* _cm;
 public:
